+++ src/frozen/__init__.py
@@ -1,5 +1,7 @@
 from .core import View
 from .freezable import freezable, freezableclass, freezablemethod, Freezable
 from .lockable import lockable, lockableclass, lockablemethod, Lockable
+from .alienatable import alienatable, alienatableclass, alienatablemethod, Alienatable
 from .freezable import FrozenError
 from .lockable import LockError, UnlockError, LockedError
+from .alienatable import AlienError